將 formatTime 的相關function移至common，使client跟server可以用共通的function來format時間
將 formatTime 的相關function(formatDateTimeText, formatShortDateTimeText, formatShortDurationTimeText, formatLongDurationTimeText, padZero)移至common下的 formatTimeUtils
使client跟server可以用共通的function來format時間
