+++ client/utils/helpers.js
@@ -12,6 +12,7 @@ import { dbEmployees } from '/db/dbEmployees';
 import { dbVariables } from '/db/dbVariables';
 import { stoneDisplayName } from '/db/dbCompanyStones';
 import { hasRole, hasAnyRoles, hasAllRoles } from '/db/users';
+import { formatDateTimeText, formatShortDateTimeText, formatShortDurationTimeText, formatLongDurationTimeText } from '/common/imports/utils/formatTimeUtils';
 
 import '../layout/highcharts-themes';
 
@@ -69,77 +70,10 @@ Template.registerHelper('getCompanyEPS', getCompanyEPS);
 Template.registerHelper('getCompanyPERatio', getCompanyPERatio);
 Template.registerHelper('getCompanyEPRatio', getCompanyEPRatio);
 
-export function formatDateTimeText(date) {
-  if (! date) {
-    return '????/??/?? ??:??:??';
-  }
 
-  return (
-    `${date.getFullYear()}/${padZero(date.getMonth() + 1)}/${padZero(date.getDate())} ${padZero(date.getHours())}:${padZero(date.getMinutes())}:${padZero(date.getSeconds())}`
-  );
-}
-function padZero(n) {
-  return n < 10 ? `0${n}` : `${n}`;
-}
 Template.registerHelper('formatDateTimeText', formatDateTimeText);
-
-export function formatShortDateTimeText(date) {
-  if (! date) {
-    return '??/?? ??:??';
-  }
-
-  return (
-    `${padZero(date.getMonth() + 1)}/${padZero(date.getDate())} ${padZero(date.getHours())}:${padZero(date.getMinutes())}`
-  );
-}
 Template.registerHelper('formatShortDateTimeText', formatShortDateTimeText);
-
-export function formatShortDurationTimeText(time) {
-  const timeBase = 1000 * 60;
-
-  if (! time) {
-    return '??:??';
-  }
-
-  time = Math.floor(time / timeBase);
-
-  return (
-    `${padZero(Math.floor(time / 60))}:${padZero(time % 60)}`
-  );
-}
 Template.registerHelper('formatShortDurationTimeText', formatShortDurationTimeText);
-
-export function formatLongDurationTimeText(time) {
-  if (! time) {
-    return '不明';
-  }
-
-  const secondBase = 1000;
-  const minuteBase = 60 * secondBase;
-  const hourBase = 60 * minuteBase;
-  const dayBase = 24 * hourBase;
-
-  let remainingTime = time;
-
-  const days = Math.floor(remainingTime / dayBase);
-  remainingTime -= days * dayBase;
-
-  const hours = Math.floor(remainingTime / hourBase);
-  remainingTime -= hours * hourBase;
-
-  const minutes = Math.floor(remainingTime / minuteBase);
-  remainingTime -= minutes * minuteBase;
-
-  const seconds = Math.floor(remainingTime / secondBase);
-  remainingTime -= seconds * secondBase;
-
-  return [
-    time >= dayBase ? `${days} 天` : '',
-    time >= hourBase ? `${hours} 時` : '',
-    time >= minuteBase ? `${minutes} 分` : '',
-    time >= secondBase ? `${seconds} 秒` : ''
-  ].join(' ').trim();
-}
 Template.registerHelper('formatLongDurationTimeText', formatLongDurationTimeText);
 
 export function currentUserId() {
